refactor: extract world topology BFS and ratchet 250-line file caps.
Split bfsTopologyGraph into a sibling library and densify leftover world
tests so lib/test/support checkers can drop from 300/280 to 250 without
changing traversal or army semantics.
