+++ SPEC/program/game-events.md
@@ -18,7 +18,7 @@ A single event stream for "what happened in the game" consumable by the Flutter 
 
 ## Event types (contract)
 
-Events are a union or sealed type (e.g. `GameEvent`) with variants. Payloads use keys/ids only; no UI strings. Province ids in any payload are **prefixed** (`regionId|localId`) per [world-model-identity.md](../game/world-model-identity.md). Implementation keeps sealed variants in one Dart library so pattern-switch exhaustiveness stays sound; multi-file splits would need `part` directives (disallowed under world package lint) or a non-sealed base (Refs #4330). That library (`packages/colonizethis_world/lib/src/game_events.dart`) has a dedicated **400** physical-line ceiling; other world lib files are gated at **300** (`repo.colonizethis_world_lib_file_size`, Refs #4515).
+Events are a union or sealed type (e.g. `GameEvent`) with variants. Payloads use keys/ids only; no UI strings. Province ids in any payload are **prefixed** (`regionId|localId`) per [world-model-identity.md](../game/world-model-identity.md). Implementation keeps sealed variants in one Dart library so pattern-switch exhaustiveness stays sound; multi-file splits would need `part` directives (disallowed under world package lint) or a non-sealed base (Refs #4330). That library (`packages/colonizethis_world/lib/src/game_events.dart`) has a dedicated **400** physical-line ceiling; other world lib files are gated at **250** (`repo.colonizethis_world_lib_file_size`, Refs #4515, #4611).
 
 | Event type           | When emitted                    | Payload (minimal) |
 |----------------------|----------------------------------|-------------------|
@@ -67,7 +67,7 @@ The **caller** that owns the order list or invokes TurnResolver is responsible f
 - **Emission — victory_set.** Given a turn resolution run in which the End-of-turn phase sets `Game.victory` (winner, type military, turn number), when the End-of-turn phase completes, then the system emits exactly one `victory_set` event with winnerPlayerId, victoryType, and turnNumber equal to the game’s victory state.
 - **Emission — order_rejected.** Given the order engine validates a player’s order list and the first rejection occurs at order N with reason R and order family K, when validation returns, then the system has emitted an `order_rejected` event for that order with `orderKind` equal to K and a `reasonCode` consistent with R (or the validation layer exposes the same so that a consumer can emit the event).
 - **Determinism.** Given the same Game (and WorldState), same per-player order lists, same ruleset, and same seeds, when the system runs turn resolution and order validation, then the sequence of GameEvents emitted for that run is identical to any other run with the same inputs.
-- **Sealed single-library size.** Given `GameEvent` remains `sealed` in `packages/colonizethis_world/lib/src/game_events.dart`, when `repo.colonizethis_world_lib_file_size` runs, then that file is allowed up to 400 physical lines and every other non-generated world lib file is gated at 300 (Refs #4515).
+- **Sealed single-library size.** Given `GameEvent` remains `sealed` in `packages/colonizethis_world/lib/src/game_events.dart`, when `repo.colonizethis_world_lib_file_size` runs, then that file is allowed up to 400 physical lines and every other non-generated world lib file is gated at 250 (Refs #4515, #4611).
 
 - **Emission — province_captured.** Given combat or another resolver step would change a province from owner **A** to owner **B**, when **A** and **B** are each a non-empty faction id and **A ≠ B**, then the system may emit a `province_captured` event with `previousOwnerId == A`, `newOwnerId == B`, and prefixed `provinceId`. Given the post-resolution `ownerId` for that province is null or empty while the previous snapshot had a non-empty owner, when the emitter runs, then the system does **not** emit `province_captured` for that province (invalid capture state; see [world-model.md](../game/world-model.md) § Invariants).
