+++ packages/colonizethis_world/test/world_test_support/sea_reachable_test_support.dart
@@ -45,7 +45,7 @@ typedef SeaReachableDistancesCase = ({
   Map<String, int> expected,
 });
 
-MapTopology _graph({
+MapTopology seaReachableTestGraph({
   required List<String> provinces,
   List<String> seas = const [],
   required List<(String, String)> edges,
@@ -62,7 +62,7 @@ final List<SeaReachableIdsCase> seaReachableIdsCases = [
   (
     description:
         'reaches a foreign province across a sea zone from an owned anchor',
-    topology: _graph(
+    topology: seaReachableTestGraph(
       provinces: const ['oldWorld|own', 'oldWorld|enemy'],
       seas: const ['oldWorld|sea'],
       edges: const [
@@ -77,7 +77,7 @@ final List<SeaReachableIdsCase> seaReachableIdsCases = [
   ),
   (
     description: 'ignores anchors the player does not actually own',
-    topology: _graph(
+    topology: seaReachableTestGraph(
       provinces: const ['oldWorld|own', 'oldWorld|enemy'],
       edges: const [('oldWorld|own', 'oldWorld|enemy')],
     ),
@@ -88,7 +88,7 @@ final List<SeaReachableIdsCase> seaReachableIdsCases = [
   ),
   (
     description: 'does not expand through a foreign province',
-    topology: _graph(
+    topology: seaReachableTestGraph(
       provinces: const ['oldWorld|own', 'oldWorld|enemy', 'oldWorld|beyond'],
       edges: const [
         ('oldWorld|own', 'oldWorld|enemy'),
@@ -106,7 +106,7 @@ final List<SeaReachableIdsCase> seaReachableIdsCases = [
   ),
   (
     description: 'skips unowned (ownerless) provinces',
-    topology: _graph(
+    topology: seaReachableTestGraph(
       provinces: const ['oldWorld|own', 'oldWorld|wild'],
       edges: const [('oldWorld|own', 'oldWorld|wild')],
     ),
@@ -117,7 +117,7 @@ final List<SeaReachableIdsCase> seaReachableIdsCases = [
   ),
   (
     description: 'regionIdFilter restricts collected foreign provinces',
-    topology: _graph(
+    topology: seaReachableTestGraph(
       provinces: const ['oldWorld|own', 'newWorld|colony', 'oldWorld|enemy'],
       seas: const ['oldWorld|sea'],
       edges: const [
@@ -134,109 +134,5 @@ final List<SeaReachableIdsCase> seaReachableIdsCases = [
     anchors: const {'oldWorld|own'},
     regionIdFilter: 'newWorld',
     expected: const {'newWorld|colony'},
-  ),
-];
-
-/// Distances-suite cases (≥320 densify, Refs #4330 Slice C).
-final List<SeaReachableDistancesCase> seaReachableDistancesCases = [
-  (
-    description:
-        'a direct province-province border foreign province has distance 1',
-    topology: _graph(
-      provinces: const ['oldWorld|own', 'oldWorld|enemy'],
-      edges: const [('oldWorld|own', 'oldWorld|enemy')],
-    ),
-    owners: const {'oldWorld|own': 'p1', 'oldWorld|enemy': 'p2'},
-    anchors: const {'oldWorld|own'},
-    regionIdFilter: null,
-    expected: const {'oldWorld|enemy': 1},
-  ),
-  (
-    description: 'counts each traversed edge, sea zone included',
-    topology: _graph(
-      provinces: const ['oldWorld|own', 'oldWorld|enemy'],
-      seas: const ['oldWorld|sea'],
-      edges: const [
-        ('oldWorld|own', 'oldWorld|sea'),
-        ('oldWorld|sea', 'oldWorld|enemy'),
-      ],
-    ),
-    owners: const {'oldWorld|own': 'p1', 'oldWorld|enemy': 'p2'},
-    anchors: const {'oldWorld|own'},
-    regionIdFilter: null,
-    expected: const {'oldWorld|enemy': 2},
-  ),
-  (
-    description: 'keeps the shortest distance when multiple paths exist',
-    topology: _graph(
-      provinces: const ['oldWorld|own', 'oldWorld|enemy'],
-      seas: const ['oldWorld|sea'],
-      edges: const [
-        ('oldWorld|own', 'oldWorld|enemy'),
-        ('oldWorld|own', 'oldWorld|sea'),
-        ('oldWorld|sea', 'oldWorld|enemy'),
-      ],
-    ),
-    owners: const {'oldWorld|own': 'p1', 'oldWorld|enemy': 'p2'},
-    anchors: const {'oldWorld|own'},
-    regionIdFilter: null,
-    expected: const {'oldWorld|enemy': 1},
-  ),
-  (
-    description: 'regionIdFilter restricts the distance map entries',
-    topology: _graph(
-      provinces: const ['oldWorld|own', 'newWorld|colony', 'oldWorld|enemy'],
-      seas: const ['oldWorld|sea'],
-      edges: const [
-        ('oldWorld|own', 'oldWorld|sea'),
-        ('oldWorld|sea', 'newWorld|colony'),
-        ('oldWorld|sea', 'oldWorld|enemy'),
-      ],
-    ),
-    owners: const {
-      'oldWorld|own': 'p1',
-      'newWorld|colony': 'p2',
-      'oldWorld|enemy': 'p2',
-    },
-    anchors: const {'oldWorld|own'},
-    regionIdFilter: 'newWorld',
-    expected: const {'newWorld|colony': 2},
-  ),
-  (
-    description:
-        'canonical NW route via owned -> OW sea -> NW sea -> NW colony is distance 3',
-    topology: _graph(
-      provinces: const ['oldWorld|own', 'newWorld|colony'],
-      seas: const ['oldWorld|owSea', 'newWorld|nwSea'],
-      edges: const [
-        ('oldWorld|own', 'oldWorld|owSea'),
-        ('oldWorld|owSea', 'newWorld|nwSea'),
-        ('newWorld|nwSea', 'newWorld|colony'),
-      ],
-    ),
-    owners: const {'oldWorld|own': 'p1', 'newWorld|colony': 'p2'},
-    anchors: const {'oldWorld|own'},
-    regionIdFilter: kNewWorldRegionId,
-    expected: const {'newWorld|colony': 3},
-  ),
-  (
-    description: 'distance BFS does not expand through foreign provinces',
-    topology: _graph(
-      provinces: const ['oldWorld|own', 'newWorld|colony', 'newWorld|far'],
-      seas: const ['newWorld|nwSea'],
-      edges: const [
-        ('oldWorld|own', 'newWorld|nwSea'),
-        ('newWorld|nwSea', 'newWorld|colony'),
-        ('newWorld|colony', 'newWorld|far'),
-      ],
-    ),
-    owners: const {
-      'oldWorld|own': 'p1',
-      'newWorld|colony': 'p2',
-      'newWorld|far': 'p3',
-    },
-    anchors: const {'oldWorld|own'},
-    regionIdFilter: kNewWorldRegionId,
-    expected: const {'newWorld|colony': 2},
   ),
 ];
