+++ tool/ct_repo_lint_manifest.yaml
@@ -965,21 +965,21 @@ rules:
 
   - rule_id: repo.colonizethis_world_lib_file_size
     group: structure
-    title: colonizethis_world lib files must stay at or below 300 physical lines except sealed GameEvent at 400 (Refs #4330, #4515)
+    title: colonizethis_world lib files must stay at or below 250 physical lines except sealed GameEvent at 400 (Refs #4330, #4515, #4611)
     spec: SPEC/program/repo-lint.md
     runner: dart
     script: tool/check_world_lib_file_size.dart
 
   - rule_id: repo.colonizethis_world_test_file_size
     group: testing
-    title: colonizethis_world non-support test files must stay at or below the wave-7 300 physical-line ceiling (Refs #4515)
+    title: colonizethis_world non-support test files must stay at or below the wave-8 250 physical-line ceiling (Refs #4515, #4611)
     spec: SPEC/program/repo-lint.md
     runner: dart
     script: tool/check_world_test_file_size.dart
 
   - rule_id: repo.world_test_support_file_size
     group: testing
-    title: colonizethis_world test/world_test_support files must stay at or below 280 physical lines (Refs #4515)
+    title: colonizethis_world test/world_test_support files must stay at or below 250 physical lines (Refs #4515, #4611)
     spec: SPEC/program/repo-lint.md
     runner: dart
     script: tool/check_world_test_support_file_size.dart
